Add d192v128 (MLA) attention kernels (forward + backward, R96C campaign)
Adds the full d_qk=192 / d_v=128 attention kernel set targeting AMD gfx950
(MI355X), bf16, GQA-aware causal:

  * attn_fwd_causal_d192v128.cpp           (forward)
  * attn_bkwd_prep_d192v128.cpp            (delta = O * dO precompute)
  * attn_bkwd_causal_d192v128_art.cpp      (combined dK/dV via ART layout)
  * attn_bkwd_dq_d192v128_art_qparallel.cpp (standalone Q-parallel dQ)

Production geometry B=16 N=16384 H=64 H_KV=8 measures 580T / 381 ms on the
combined backward — +13% over AMD's hand-written FA-3 v3 ASM
(bwd_hd192_128_bf16_causal_br_a32_pssk.co at 513T / 428 ms).

Required framework changes (load-bearing for the kernels above):

  * include/common/macros.cuh                              — DS load/store +
    16x16x32 / 32x32x16 bf16 MFMA wrappers
  * include/ops/warp/memory/tile/global_to_shared.cuh      — swizzle-aware
    G::load (R98-γ) + 64-bit buffer-descriptor clamp (R90B)
  * include/ops/warp/memory/tile/shared_to_register.cuh    — pad-aware load
    paths used by attn_smem (R94G)
  * include/ops/warp/memory/tile/assembly/shared_to_register.cuh
  * include/ops/warp/memory/vec/global_to_shared.cuh
  * include/ops/warp/register/tile/conversions.cuh         — bf16 cvt forms
    used in the dKdV cvt-into-MFMA-shadow hoist (R96C)
  * include/ops/warp/register/tile/assembly/conversions.cuh
  * include/ops/warp/register/tile/mma.cuh + assembly/mma.cuh
  * include/types/global/gl.cuh
  * include/types/register/rt_shape.cuh
  * include/types/shared/st.cuh + st_shape.cuh             — pad{1,2,4}
    swizzle shapes
  * include/types/types.cuh

Build:
  cd kernels/attn/gqa_causal_backwards && make asymmetric ATTN_D_QK=192

Bench:
  python bench_full.py

Campaign closeout (decisions, lever-class auto-rejects, profiling notes)
captured in TODO.md and agent_prompt.md alongside the kernels.

+++ include/types/global/gl.cuh
@@ -69,10 +69,12 @@ struct gl {
     __host__ __device__ inline gl(const gl &other) :
             raw_ptr(other.raw_ptr), batch_internal(other.batch_internal), depth_internal(other.depth_internal), rows_internal(other.rows_internal), cols_internal(other.cols_internal), tma_descs(other.tma_descs) {}
     __device__ inline T& operator[](const coord<ducks::default_type> &idx) const { // yes I am abusing the const qualifier here a bit.
-        return raw_ptr[((idx.b*depth() + idx.d)*rows() + idx.r)*cols() + idx.c];
+        // R23-A: cast to size_t to avoid int32 overflow when total tensor > 2 GiB
+        // (e.g., production-shape FP8 attention P[B=16,H=64,N=4096,N=4096] = 16 GiB).
+        return raw_ptr[((size_t(idx.b)*depth() + idx.d)*rows() + idx.r)*cols() + idx.c];
     }
-    __device__ inline int idx(const coord<ducks::default_type> &idx) const {
-        return ((idx.b*depth() + idx.d)*rows() + idx.r)*cols() + idx.c;
+    __device__ inline size_t idx(const coord<ducks::default_type> &idx) const {
+        return ((size_t(idx.b)*depth() + idx.d)*rows() + idx.r)*cols() + idx.c;
     }
     template<int axis> __device__ inline size_t shape() const {
         static_assert(axis==0 || axis==1 || axis==2 || axis==3, "Axis must be 0, 1, 2, or 3.");
